refactor(geometry): split the rep-identity grouping pass out of collate.rs
collate.rs reached 475 lines once this round's refusal handling landed on
top of the author's dont_bake split, past the 400-line rule with no
allowlist row. collate.rs keeps the collation types and the small frame
helpers; group.rs owns the pass that walks the groups, verifies each
candidate pairing (#3666) and decides what falls back to flat.
